\name{Bpkg-package}
\alias{Bpkg-package}
\alias{Bpkg}
\docType{package}
\title{
\packageTitle{Bpkg}
}
\description{
\packageDescription{Bpkg}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Bpkg}

Maintainer: \packageMaintainer{Bpkg}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Bpkg)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Bpkg}}
%\section{Documentation Index}{\packageIndices{Bpkg}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
