% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgB.R
\docType{package}
\name{pkgB-package}
\alias{pkgB-package}
\title{pkgB that links to native routines in pkgA}
\description{
pkgB that links to native routines in pkgA
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
