% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgB.R
\name{pkgBfunc}
\alias{pkgBfunc}
\title{A dummy function that shows linking to native routines in pkgA works}
\usage{
pkgBfunc()
}
\value{
Side effect is used
}
\description{
A dummy function that shows linking to native routines in pkgA works
}
